🐛(lib-video) intl not accessible from toast.custom
When we load a component from a toast.custom, we can't access the
intl object, the ancestor tree is lost. We need to pass it down
explicitly.
